// admin/src/domains/inventory/bundles/modals/dependency-chart-modal/types.ts
export interface SelectedRuleNode {
  type: "rule";
  rule: IDependencyRule;
}

export interface SelectedItemNode {
  type: "item";
  item: BundleItem;
  group: IBundleGroup;
}

export interface SelectedGroupNode {
  type: "group";
  group: IBundleGroup;
}

export interface SelectedBundleNode {
  type: "bundle";
  label: string;
}

export type SelectedNode =
  | SelectedRuleNode
  | SelectedItemNode
  | SelectedGroupNode
  | SelectedBundleNode
  | null;